refactor(file): remove filecontroller and route
