+++ apps/api/src/server.ts
@@ -31,7 +31,6 @@ import collaboratorsRouter from '@/routes/collaboratorsRouter.js';
 import { serverConfig } from './config/serverConfig.js';
 import logger from './logger.js';
 import authRouter from './routes/authRouter.js';
-import fileRouter from './routes/fileRouter.js';
 import sessionMiddleware from './session/sessionMiddleware.js';
 
 const __filename = fileURLToPath(import.meta.url);
@@ -49,7 +48,6 @@ const startServer = async () => {
 	app.use(applicationRouter);
 	app.use(collaboratorsRouter);
 	app.use('/auth', authRouter);
-	app.use('/file', fileRouter);
 
 	app.use(
 		`/${API_DOCS_PATH}`,
